refactor: #432 modelo dependencia, aparte
