+++ src/app/@core/data/models/oikos.ts
@@ -1,0 +1,11 @@
+/**
+ * Estructuras de datos asociadas a Oikos
+ */
+
+import { Base } from './base'
+
+export class Dependencia extends Base {
+    Nombre: string;
+    TelefonoDependencia: string;
+    CorreoElectronico: string;
+}
