+++ src/app/@core/data/models/terceros_criterio.ts
@@ -1,13 +1,13 @@
 import { EspacioFisico } from './ubicacion/espacio_fisico';
 import { Tercero } from './terceros';
-import { Dependencia } from './acta_recibido/soporte_acta';
+import { Dependencia } from './oikos';
 
 /**
  * Estructura retornada por
  * ARKA_SERVICE/terceros/tipo/funcionarioPlanta
  */
 export class TerceroCriterioPlanta {
-    Dependencia: Dependencia;
+    Dependencia: Partial<Dependencia>;
     DependenciaId: number;
     Sede: EspacioFisico;
     TerceroPrincipal: Tercero;
