+++ src/app/@core/store/app.state.ts
@@ -1,6 +1,7 @@
 import { Proveedor } from '../data/models/acta_recibido/Proveedor';
 import { Cuenta } from '../data/models/catalogo/cuenta_contable';
-import { Ubicacion, Dependencia } from '../data/models/acta_recibido/soporte_acta';
+import { Ubicacion } from '../data/models/acta_recibido/soporte_acta';
+import { Dependencia } from '../data/models/oikos'
 import { EstadoActa } from '../data/models/acta_recibido/estado_acta';
 import { EstadoElemento } from '../data/models/acta_recibido/estado_elemento';
 import { TipoBien } from '../data/models/acta_recibido/tipo_bien';
