+++ src/app/pages/acta-recibido/registro-acta-recibido/registro-acta-recibido.component.ts
@@ -11,7 +11,7 @@ import { ActaRecibidoHelper } from '../../../helpers/acta_recibido/actaRecibidoH
 import { ActaRecibido } from '../../../@core/data/models/acta_recibido/acta_recibido';
 import { Elemento, Impuesto } from '../../../@core/data/models/acta_recibido/elemento';
 import { TipoBien } from '../../../@core/data/models/acta_recibido/tipo_bien';
-import { SoporteActa, Ubicacion, Dependencia } from '../../../@core/data/models/acta_recibido/soporte_acta';
+import { SoporteActa, Ubicacion } from '../../../@core/data/models/acta_recibido/soporte_acta';
 import { ActaRecibidoUbicacion } from '../../../@core/data/models/acta_recibido/acta_recibido';
 
 import { Proveedor } from '../../../@core/data/models/acta_recibido/Proveedor';
